refactor: move block-drag-controller to activity folder
