+++ js/__tests__/blocks.test.js
@@ -78,7 +78,8 @@ global.ConnectionValidator = require("../connection-validator");
 
 // Use the real BlockDragController so drag-group and dock-snapping
 // behavior stays accurate rather than silently becoming a no-op.
-global.setupBlockDragController = require("../block-drag-controller").setupBlockDragController;
+global.setupBlockDragController =
+    require("../activity/block-drag-controller").setupBlockDragController;
 
 // Mock Constants
 global.DEFAULTBLOCKSCALE = 1.0;
