+++ js/activity/__tests__/block-drag-controller.test.js
