+++ js/activity/block-drag-controller.js
