+++ js/blocks/__tests__/RhythmBlocks.test.js
@@ -24,7 +24,8 @@ const { setupRhythmBlocks } = jest.requireActual("../RhythmBlocks");
 
 // Use the real BlockDragController so the Blocks constructor can attach
 // drag delegation methods (blocks.js calls setupBlockDragController(this)).
-global.setupBlockDragController = require("../../block-drag-controller").setupBlockDragController;
+global.setupBlockDragController =
+    require("../../activity/block-drag-controller").setupBlockDragController;
 
 const Blocks = jest.requireActual("../../blocks");
