+++ js/loader.js
@@ -232,6 +232,7 @@ requirejs.config({
         "activity/workspace-layout-controller": "js/activity/workspace-layout-controller",
         "activity/selection-controller": "js/activity/selection-controller",
         "activity/block-scale-controller": "js/activity/block-scale-controller",
+        "activity/block-drag-controller": "js/activity/block-drag-controller",
         "search-ui": "js/search-ui",
         "project-manager": "js/project-manager",
         "keyboard-controller": "js/keyboard-controller",
